+++ app/shared/infra/postgres/__init__.py
@@ -1,0 +1,10 @@
+from app.shared.infra.postgres.component import SessionMaker, SqlAlchemyComponent
+from app.shared.infra.postgres.dependency import db_session_dependency
+from app.shared.infra.postgres.uow import BaseUnitOfWork
+
+__all__ = [
+    "BaseUnitOfWork",
+    "SessionMaker",
+    "SqlAlchemyComponent",
+    "db_session_dependency",
+]