+++ app/shared/schemas/base.py
@@ -4,8 +4,8 @@ from typing import TYPE_CHECKING
 from fastapi import FastAPI
 
 if TYPE_CHECKING:
-    from app.shared.infra.components.base import BaseComponent
-    from app.shared.infra.components.registry import ComponentRegistry
+    from app.shared.infra.di.base import BaseComponent
+    from app.shared.infra.di.registry import ComponentRegistry
 
 
 class BFastAPI(FastAPI):