+++ app/shared/utils/__init__.py
@@ -1,23 +1,5 @@
-from app.shared.utils.event_names import get_event_name
 from app.shared.utils.file_reader import read_file
-from app.shared.utils.http_logging import (
-    build_error_log_context,
-    build_log_context,
-    extract_request_context,
-    get_log_level_for_exception,
-    get_status_code_from_exception,
-)
-from app.shared.utils.json_serializer import serialize_to_json
-from app.shared.utils.logger import get_logger
 
 __all__ = [
-    "build_error_log_context",
-    "build_log_context",
-    "extract_request_context",
-    "get_event_name",
-    "get_log_level_for_exception",
-    "get_logger",
-    "get_status_code_from_exception",
     "read_file",
-    "serialize_to_json",
 ]